Desugar pairs with no namespace
Two small fixes were necassary in the ambiguous resolution for this to
work, making sure Metadata was changed to the successful elaboration
result, and not increasing the ambiguity tries when falling back to a
unique default in alternatives.

Data.Container in contrib also has had it's Pair constructor name
changed to avoid ambiguities, as it was using both that and
Builtin.Pair.

Fixes #634.